+++ connectToAWSServerViaSSH.sh
@@ -1,1 +1,1 @@
-ssh -i ~/.ssh/linux-demo.pem ec2-user@3.16.123.110
+ssh -i ~/.ssh/linux-demo.pem ec2-user@3.17.71.66