+++ src/main/java/com/github/jeffmagina/project1/ServerApp.java
@@ -19,12 +19,12 @@ public class ServerApp {
 	
 	public static void main(String args[]) {
 		
-		//runs server
+		// run server
 		if (args[0].equalsIgnoreCase("server")) {
 			startTomcat();
 		} 
 		
-		// used to upload text file data and run server
+		// upload text file data and run server
 		else if (args[0].equalsIgnoreCase("loadCSV")) {
 			String textFile = args[1];
 			
@@ -37,38 +37,8 @@ public class ServerApp {
 			//loadtext file
 			JavaRDD<MyCSVFile> data = new LoadCSVFile(textFile).createRDD(sc);
 			
-			//spark manipulation
+			//send to to database			
 			SparkTransformations sparkTransformations = new SparkTransformations(data);
-			
-			System.out.println(sparkTransformations.getDataStorage("countGender"));
-			
-			System.out.println(sparkTransformations.getDataStorage("countRace"));
-			
-			System.out.println(sparkTransformations.getDataStorage("countParentEducation"));
-			
-			System.out.println(sparkTransformations.getDataStorage("countLunch"));
-			
-			System.out.println(sparkTransformations.getDataStorage("countTestPrep"));
-			
-			System.out.println("Math Score Avg is: " + sparkTransformations.getDataStorage("mathScoreAvg"));
-
-			System.out.println("Reading Score Avg is: " + sparkTransformations.getDataStorage("readingScoreAvg"));
-
-			System.out.println("Writing Score Avg is: " + sparkTransformations.getDataStorage("writingScoreAvg"));
-			
-			System.out.println("Overall Score Avg is: " + sparkTransformations.getDataStorage("overallScoreAvg"));
-			
-			System.out.println("Overall Gender Score Avg is: " + sparkTransformations.getDataStorage("genderAvg"));
-			
-			System.out.println("Overall Race Score Avg is: " + sparkTransformations.getDataStorage("raceAvg"));
-			
-			System.out.println("Overall ParentEducation Score Avg is: " + sparkTransformations.getDataStorage("parentEducationAvg"));
-			
-			System.out.println("Overall Lunch Score Avg is: " + sparkTransformations.getDataStorage("lunchAvg"));
-			
-			System.out.println("Overall TestPrep Score Avg is: " + sparkTransformations.getDataStorage("testPrepAvg"));
-			
-			//send to to database
 			LinkedHashMap<String, String> dataStorage = sparkTransformations.getDataStorage();
 			SQLRepo sqlRepo = new SQLRepo();
 			
@@ -78,9 +48,6 @@ public class ServerApp {
 
 			//initiate embed tomcat
 			startTomcat();
-			
-			//close JavaSparkContect
-			//sc.close();
 		}
 	}